Extended spring security implementation

+++ src/main/java/service/impl/assesment/AssessmentManagerImpl.java
@@ -100,11 +100,11 @@ public class AssessmentManagerImpl implements AssessmentManager
                 process = createAssessmentProcess();
                 process.setAssessment( assessment );
                 process.setUser( user );
-                process.setState( ProcessState.Ready.getId() );
             }
             
             if(process.getState() != ProcessState.Finished.getId() )
             {
+                process.setState( ProcessState.Ready.getId() );
                 process.setTaskIds( taskDAO.getRandomIdByAssessmentId( assessment.getId() ) );
             }
             
@@ -124,22 +124,31 @@ public class AssessmentManagerImpl implements AssessmentManager
         
         try
         {
+            //-----------Check Valid ----------------------------
             if(process.getState() == ProcessState.Finished.getId())
             {
                 throw new TimeExpiredException("Process already finished (Test passed):"+ process.getAssessment().getName());
             }
             
-            if(isTimeExpired( process.getAssessment().getTime(), process.getEndDate() ))
+            if(isTimeExpired( process.getAssessment().getTime(), process.getStartDate() ))
             {
                 //-----End Process if time expired -------------
                 endProcess(process, null);
                 //----------------------------------------------
                 throw new TimeExpiredException("Time is expired for Assessment:"+ process.getAssessment().getName());
             }
-
-            //*******************************************************
+            //----------------------------------------------
+            
             if(process.getState() == ProcessState.Ready.getId() )
             {
+                process.setStartDate( new Date(System.currentTimeMillis()) );
+                process.setEndDate( process.getStartDate() );
+                process.setState( ProcessState.Started.getId() );
+                processDAO.save( process );
+            }
+            else if(process.getState() == ProcessState.Resumed.getId() )
+            {
+                process.setEndDate( new Date(System.currentTimeMillis()) );
                 process.setState( ProcessState.Started.getId() );
                 processDAO.save( process );
             }
@@ -151,7 +160,10 @@ public class AssessmentManagerImpl implements AssessmentManager
                 if(!CollectionUtils.isEmpty( processResponse.getDetails() ))
                 {
                     // ----------Remove previous responses -----------
-                    processResponseDAO.delete( processResponse.getId() );
+                    if(processResponse.getId() > 0)
+                    {
+                        processResponseDAO.delete( processResponse.getId() );
+                    }
                     // -----------------------------------------------
                     
                     Iterator<ProcessResponseDetail> itr = processResponse.getDetails().iterator();
@@ -285,7 +297,10 @@ public class AssessmentManagerImpl implements AssessmentManager
      */
     private boolean isTimeExpired( int time, Date lastActiveDate )
     {
-        lastActiveDate = DateUtils.addMinutes( lastActiveDate, time );
+        if(lastActiveDate == null)
+            return false;
+        
+        lastActiveDate = DateUtils.addSeconds( lastActiveDate, (time * 60) );
         Date currentDate = new Date(System.currentTimeMillis());
         
         if( lastActiveDate.getTime() < currentDate.getTime())
